02.Serverless-Automation-with-GCP-and-Shelly-Devices/resources/pubsub-to-mqtt/requirements.txt
functions-framework==3.*
paho-mqtt